(trunk libT) #3767 "rarest first policy" -- fixed.
This commit, started by a patch from athy, implements a rarest first policy when deciding which pieces to request from peers. It keeps a count of how many peers have each piece, and updates the count when getting bitfields, have, have all, and have none messages, as well as decrementing the counts when peers disconnect.

This running total is generated only for downloading torrents. Seeds don't have this overhead.


git-svn-id: svn://svn.transmissionbt.com/Transmission/trunk@11897 f4695dd4-2c0a-0410-b89c-da849a56a58e